staging_xcm_executor

Module traits

Various traits used in configuring the executor.

Modules

prelude

Structs

Identity

A structure that performs identity conversion.

JustTry

A structure that performs standard conversion using the standard Rust conversion traits.

Properties

Properties of an XCM message and its imminent execution.

WithOriginFilter

Enums

AssetTransferError

Errors related to determining asset transfer support.

Error

Errors associated with MatchesFungibles operation.

FeeReason

Context under which a fee is paid.

LockError

QueryResponseStatus

The possible state of an XCM query response.

TransferType

Specify which type of asset transfer is required for a particular (asset, dest) combination.

Traits

AssetExchange

A service for exchanging assets.

AssetLock

Define a handler for notification of an asset being locked and for the unlock instruction.

CallDispatcher

Defines how a call is dispatched with given origin. Allows to customize call dispatch, such as adapting the origin based on the call or modifying the call.

CheckSuspension

Trait to determine whether the execution engine is suspended from executing a given XCM.

ClaimAssets

Define any handlers for the AssetClaim instruction.

ConvertLocation

Means of converting a location into an account identifier.

ConvertOrigin

A converter trait for origin types.

DenyExecution

Trait to determine whether the execution engine should not execute a given XCM.

DropAssets

Define a handler for when some non-empty AssetsInHolding value should be dropped.

Enact

EventEmitter

Defines the event emitter for the XCM executor. This trait allows implementations to emit events related to XCM handling, including successful sends and failure cases.

ExportXcm

Utility for delivering a message to a system under a different (non-local) consensus with a spoofed origin. This essentially defines the behaviour of the ExportMessage XCM instruction.

FeeManager

Handle stuff to do with taking fees in certain XCM instructions.

FilterAssetLocationDeprecated

Filters assets/location pairs.

HandleHrmpChannelAccepted

Executes optional logic when a HrmpChannelAccepted XCM notification is received.

HandleHrmpChannelClosing

Executes optional logic when a HrmpChannelClosing XCM notification is received.

HandleHrmpNewChannelOpenRequest

Executes logic when a HrmpNewChannelOpenRequest XCM notification is received.

MatchesFungible

MatchesFungibles

MatchesNonFungible

MatchesNonFungibles

OnResponse

Define what needs to be done upon receiving a query response.

ProcessTransaction

Provides mechanisms for transactional processing of XCM instructions.

QueryHandler

Provides methods to expect responses from XCMs and query their status.

RecordXcm

Trait for recording XCMs.

ShouldExecute

Trait to determine whether the execution engine should actually execute a given XCM.

TransactAsset

Facility for asset transacting.

VersionChangeNotifier

Trait for a type which handles notifying a destination of XCM version changes.

WeightBounds

Determine the weight of an XCM message.

WeightTrader

Charge for weight in order to execute XCM.

XcmAssetTransfers

A trait for identifying asset transfer type based on IsTeleporter and IsReserve configurations.

Functions

export_xcm

Convenience function for using a SendXcm implementation. Just interprets the dest and wraps both in Some before passing them as mutable references into T::send_xcm.

validate_export

Convenience function for using a SendXcm implementation. Just interprets the dest and wraps both in Some before passing them as mutable references into T::send_xcm.