Trait staging_xcm_executor::traits::CallDispatcher

pub trait CallDispatcher<Call: Dispatchable> {
    // Required method
    fn dispatch(
        call: Call,
        origin: Call::RuntimeOrigin,
    ) -> Result<Call::PostInfo, DispatchErrorWithPostInfo<Call::PostInfo>>;
}

Defines how a call is dispatched with given origin. Allows to customize call dispatch, such as adapting the origin based on the call or modifying the call.

Required Methods

fn dispatch( call: Call, origin: Call::RuntimeOrigin, ) -> Result<Call::PostInfo, DispatchErrorWithPostInfo<Call::PostInfo>>

Object Safety

This trait is not object safe.