staging_xcm_executor::traits

Trait FeeManager

pub trait FeeManager {
    // Required methods
    fn is_waived(origin: Option<&Location>, r: FeeReason) -> bool;
    fn handle_fee(fee: Assets, context: Option<&XcmContext>, r: FeeReason);
}

Handle stuff to do with taking fees in certain XCM instructions.

Required Methods

fn is_waived(origin: Option<&Location>, r: FeeReason) -> bool

Determine if a fee should be waived.

fn handle_fee(fee: Assets, context: Option<&XcmContext>, r: FeeReason)

Do something with the fee which has been paid. Doing nothing here silently burns the fees.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.