Trait staging_xcm_executor::traits::OnResponse

pub trait OnResponse {
    // Required methods
    fn expecting_response(
        origin: &Location,
        query_id: u64,
        querier: Option<&Location>,
    ) -> bool;
    fn on_response(
        origin: &Location,
        query_id: u64,
        querier: Option<&Location>,
        response: Response,
        max_weight: Weight,
        context: &XcmContext,
    ) -> Weight;
}

Define what needs to be done upon receiving a query response.

Required Methods

fn expecting_response( origin: &Location, query_id: u64, querier: Option<&Location>, ) -> bool

Returns true if we are expecting a response from origin for query query_id that was queried by querier.

fn on_response( origin: &Location, query_id: u64, querier: Option<&Location>, response: Response, max_weight: Weight, context: &XcmContext, ) -> Weight

Handler for receiving a response from origin relating to query_id initiated by querier.

Object Safety

This trait is not object safe.