Trait staging_xcm_executor::traits::XcmAssetTransfers

pub trait XcmAssetTransfers {
    type IsReserve: ContainsPair<MultiAsset, MultiLocation>;
    type IsTeleporter: ContainsPair<MultiAsset, MultiLocation>;
    type AssetTransactor: TransactAsset;

    // Provided method
    fn determine_for(
        asset: &MultiAsset,
        dest: &MultiLocation
    ) -> Result<TransferType, Error> { ... }
}

A trait for identifying asset transfer type based on IsTeleporter and IsReserve configurations.

Required Associated Types

type IsReserve: ContainsPair<MultiAsset, MultiLocation>

Combinations of (Asset, Location) pairs which we trust as reserves. Meaning reserve-based-transfers are to be used for assets matching this filter.

type IsTeleporter: ContainsPair<MultiAsset, MultiLocation>

Combinations of (Asset, Location) pairs which we trust as teleporters. Meaning teleports are to be used for assets matching this filter.

type AssetTransactor: TransactAsset

How to withdraw and deposit an asset.

Provided Methods

fn determine_for( asset: &MultiAsset, dest: &MultiLocation ) -> Result<TransferType, Error>

Determine transfer type to be used for transferring asset from local chain to dest.