Function staging_xcm_executor::traits::validate_export

pub fn validate_export<T: ExportXcm>(
    network: NetworkId,
    channel: u32,
    universal_source: InteriorLocation,
    dest: InteriorLocation,
    msg: Xcm<()>,
) -> SendResult<T::Ticket>

Convenience function for using a SendXcm implementation. Just interprets the dest and wraps both in Some before passing them as as mutable references into T::send_xcm.